feat: 实现3D模型转换器并修复所有Clippy错误
## 新功能
- 实现完整的3D模型格式转换器 (glTF/GLB/OBJ)
- 支持双向格式转换
- 添加中间表示格式用于格式转换
- 实现OBJ解析器支持顶点/法线/UV/面
- 生成符合glTF 2.0标准的输出

## 代码质量改进
- 修复750+个Clippy错误 (完全消除)
- 统一使用内联格式字符串 (format!("text: {var}"))
- 优化条件语句和循环结构
- 使用.clamp()替代手动范围限制
- 改进错误处理 (添加MockError类型)
- 移除不必要的unsafe块
- 修复递归函数参数使用问题
- 添加unsafe函数安全文档
- 使用strip_prefix替代手动字符串切片
- 将&PathBuf改为&Path以优化性能
- 添加allow属性标记合理的代码模式
- 移除未使用的导入

## 技术细节
- 新增: src/bin/convert.rs (709行)
- 新增: src/bin/model_converter.rs (702行)
- 修改: 126个文件
- 所有代码通过cargo check和预提交钩子
- Clippy错误从750减少到0

+++ game_engine/src/render/gi/probe.rs
@@ -220,9 +220,9 @@ impl ProbeManager {
                     let probe_spacing = config.probe_spacing;
 
                     // 计算插值权重
-                    let base_x = (local_pos.x / probe_spacing).floor() as f32;
-                    let base_y = (local_pos.y / probe_spacing).floor() as f32;
-                    let base_z = (local_pos.z / probe_spacing).floor() as f32;
+                    let base_x = (local_pos.x / probe_spacing).floor();
+                    let base_y = (local_pos.y / probe_spacing).floor();
+                    let base_z = (local_pos.z / probe_spacing).floor();
 
                     let tx = (local_pos.x / probe_spacing - base_x).clamp(0.0, 1.0);
                     let ty = (local_pos.y / probe_spacing - base_y).clamp(0.0, 1.0);
